moved formatName() from library/remote-store.ts to utils.ts, formatStringList() from library/log-functions to utils.ts, toYMD() from library/calendar.ts and app/(modals)/calendar.tsx to utils.ts, safeDecrypt from library/log-functions.ts and library/calendar.ts to library/crypto.ts Added new files library/utils.ts and test/library/utils.ts
