+++ app/(modals)/active-child.tsx
@@ -10,7 +10,8 @@ import { router } from 'expo-router';
 import { SafeAreaView } from 'react-native-safe-area-context';
 import Button from '@/components/button';
 import RenameChildPopup from '@/components/rename-child-popup';
-import { formatName, getActiveChildData, updateChildName, getChildren, deleteChild } from '@/library/remote-store';
+import { getActiveChildData, updateChildName, getChildren, deleteChild } from '@/library/remote-store';
+import { formatName } from '@/library/utils';
 import supabase from '@/library/supabase-client';
 import SwitchChildPopup from '@/components/switch-child-popup';
 import stringLib from "@/assets/stringLibrary.json";
