+++ library/auth-provider.tsx
@@ -8,7 +8,7 @@ import {
     useState,
 } from 'react';
 import supabase from '@/library/supabase-client';
-import { formatName } from './remote-store';
+import { formatName } from '@/library/utils';
 import { enterGuestMode, exitGuestMode, isGuestMode } from '@/library/local-store';
 
 const AuthContext = createContext<AuthContextType | undefined>(undefined);
