+++ library/local-store.ts
@@ -1,7 +1,7 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import * as Crypto from "expo-crypto";
 import { encryptData, decryptData } from "./crypto";
-import { formatName } from "./remote-store";
+import { formatName } from "@/library/utils";
 
 export type TableName =
 	| "feeding_logs"
